# docs/v1_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Booking 2gis microservice.",
        "title": "Booking Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/v1/orders/{order_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "name": "order_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorDetailComponentResponse": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field that caused the error (if any)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "source"
                    ]
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "source: Field valid values are [foo bar] (enum)"
                }
            }
        },
        "ErrorDetailsResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Details of the error",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ErrorDetailComponentResponse"
                    }
                },
                "message": {
                    "description": "Detailed message of the error",
                    "type": "string",
                    "example": "Validation failed for the request parameters"
                },
                "slug": {
                    "description": "Slug of the error",
                    "type": "string",
                    "example": "400_malformed_request"
                },
                "status": {
                    "description": "Small message of the error",
                    "type": "string",
                    "example": "Unprocessable Entity"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error response",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ErrorDetailsResponse"
                        }
                    ]
                }
            }
        },
...
        "OrderResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "foo@bar.com"
                },
                "from": {
                    "type": "string",
                    "format": "date",
                    "example": "2021-01-01"
                },
                "hotel_id": {
                    "type": "string",
                    "example": "resort"
                },
                "id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string",
                    "example": "premium"
                },
                "to": {
                    "type": "string",
                    "format": "date",
                    "example": "2021-01-02"
                }
            }
        }
    }
}